(PUP-3900) Rename puppetx.rb to puppet_x.rb
To stay consistent with the rename of Puppetx to PuppetX and puppetx
to puppet_x, rename lib/puppetx.rb to lib/puppet_x.rb and update some
corresponding tests.
