+++ lib/puppet_x.rb
